Remove Castclass from complex constants on .NET Core
It seems that this was introduced to mitigate an issue on .NET Framework,
so I left it on this platform. However, it seems unnecessary on
.NET Core and seems to create slight performance trouble
for me.

The NETFRAMEWORK constnant should be automatically defined by C#,
so no need to define it in csproj
(https://docs.microsoft.com/en-us/dotnet/standard/frameworks#how-to-specify-a-target-framework)